requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
breadability==0.1.20
cachetools==6.2.2
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
docopt==0.6.2
Flask==3.1.2
google-auth==2.43.0
google-genai==1.52.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
nltk==3.9.2
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.12.4
pydantic_core==2.41.5
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sumy==0.11.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
